+++ packages/core/utils/dependencies.ts
@@ -6,7 +6,9 @@ import { executeCli } from "./common";
 export async function suggestInstallingDependencies(workingDirectory: string) {
     type PackageManager = keyof typeof packageManagers | undefined;
     const packageManagers = {
-        npm: "npm install",
+        // Prevents npm from crashing if the peer-deps don't match.
+        npm: "npm install --legacy-peer-deps",
+        // All of the other PMs should handle this more gracefully by default
         pnpm: "pnpm install",
         yarn: "yarn",
         bun: "bun install",