+++ packages/testing-library/utils/workspace.ts
@@ -12,7 +12,9 @@ export function getTemplatesDirectory(options: TestOptions) {
 
 export async function installDependencies(output: string) {
     try {
-        await executeCli("npm", ["install"], output, { stdio: "pipe" });
+        // Since tests are executed and installed within this repo (packages/tests/.outputs),
+        // we need to add the `--ignore-workspace` flag so that our root lockfile isn't modified
+        await executeCli("pnpm", ["install", "--ignore-workspace"], output, { stdio: "pipe" });
     } catch (error) {
         const errorString = error as string;
         throw new Error("unable to install dependencies: " + errorString);