+++ adders/storybook/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @svelte-add/storybook
 
+## 1.0.11
+
+### Patch Changes
+
+-   fix: move storybook from styling category to tools category ([#395](https://github.com/svelte-add/svelte-add/pull/395))
+
+-   Updated dependencies [[`c8d0622442ada1388d2f96ef9fe81c27a2f7e5eb`](https://github.com/svelte-add/svelte-add/commit/c8d0622442ada1388d2f96ef9fe81c27a2f7e5eb), [`c8d0622442ada1388d2f96ef9fe81c27a2f7e5eb`](https://github.com/svelte-add/svelte-add/commit/c8d0622442ada1388d2f96ef9fe81c27a2f7e5eb), [`c8d0622442ada1388d2f96ef9fe81c27a2f7e5eb`](https://github.com/svelte-add/svelte-add/commit/c8d0622442ada1388d2f96ef9fe81c27a2f7e5eb), [`9a116d4057407b15a99258541e54eaee6c8b7f1b`](https://github.com/svelte-add/svelte-add/commit/9a116d4057407b15a99258541e54eaee6c8b7f1b), [`db58988066790b2cd79fb5325758b87f92ce9565`](https://github.com/svelte-add/svelte-add/commit/db58988066790b2cd79fb5325758b87f92ce9565)]:
+    -   @svelte-add/core@2.5.0
+
 ## 1.0.10
 
 ### Patch Changes